+++ api/functions/deleteActivity/index.ts
@@ -1,9 +1,8 @@
 import { app, type HttpRequest, type HttpResponseInit } from "@azure/functions";
-import { getUserId } from "../authorization/firebaseAuthorization";
-import { firebaseDB as database } from "../database/firebaseDB";
-import type { ActivityRecord } from "../utils/types";
+import { getUserId } from "../../authorization/firebaseAuthorization";
+import { firebaseDB as database } from "../../database/firebaseDB";
 
-async function editActivity(request: HttpRequest): Promise<HttpResponseInit> {
+async function deleteActivity(request: HttpRequest): Promise<HttpResponseInit> {
   const idToken = request.headers.get("x-auth-token");
   let userId: string;
   try {
@@ -12,7 +11,6 @@ async function editActivity(request: HttpRequest): Promise<HttpResponseInit> {
     return { status: 401, body: (err as Error).message };
   }
 
-  const activity = (await request.json()) as ActivityRecord;
   const activityId = request.params.activityId;
 
   if (!activityId) {
@@ -20,16 +18,16 @@ async function editActivity(request: HttpRequest): Promise<HttpResponseInit> {
   }
 
   try {
-    await database.editActivity(userId, activityId, activity);
+    await database.deleteActivity(userId, activityId);
     return { status: 204 };
   } catch (err) {
     return { status: 404, body: (err as Error).message };
   }
 }
 
-app.http("editActivity", {
-  methods: ["PUT"],
+app.http("deleteActivity", {
+  methods: ["DELETE"],
   authLevel: "anonymous",
   route: "activities/{activityId}",
-  handler: editActivity,
+  handler: deleteActivity,
 });
